fix(await): fix \`isSet\` implementation
affects: @qio/core

issue: \`isSet\` when called aafter \`set\` would always return true. This is not very useful because we
don't know when the IO is resolved. This change makes sure that \`isSet\` will return false unless the
IO actually resolves. To know if the provided IO will be set can be inferred from the return value
of \`isSet\`.